Devops: fix the nightly build
The nightly build against the `develop` branch of `aiida-core` was
failing at the installation of the Python dependencies when running
against Python 3.7. The reason is that the current `develop` branch has
dropped support for Python 3.7 and the `pip` is then failing to install
a version of `pymatgen` that matches the dependency requirements. The
solution is simply to not run against Python 3.7 anymore.

Furthermore, the workflow is updated to match the exact same setup
procedure as the `ci.yml` workflow which has also recently been updated.
The workflow is also renamed to `develop` as that is the name of the
file itself and is a bit more concrete than nightly build, since you can
have many nightly builds of different kinds in principle, but this one
is especially used to test against `aiida-core`'s `develop` branch.